Handle Spotlight search and qucik menu
